Blog Website
Project Name: Blog Website

Description:
This project is a responsive blog website built using HTML, CSS, and JavaScript. It provides a platform for users to read and blog posts on various topics. The website features a clean and modern design with a user-friendly interface.

Key Features:
- Responsive design: The website is optimized for different screen sizes, ensuring a seamless user experience across devices.
- Blog post listing: Display a list of blog posts with NewsFeed, Education, Jobs and many more.
- Blog post details: Show the full content of a selected blog post with comments.

Technologies Used:
- HTML: Markup language used for structuring the web pages.
- CSS: Stylesheet language used for designing the website layout and appearance.
- JavaScript: Programming language used to add interactivity and dynamic functionality.

Deployment:
The website can be deployed on any web server or hosting platform supporting HTML, CSS, and JavaScript. You can simply clone the repository and open the index.html file in a web browser to view the website locally.

Contributions:
Contributions to the project are welcome! If you find any issues or have suggestions for improvements, feel free to open an issue or submit a pull request on GitHub.

License:
Specify the license under which the project is distributed, such as General Public License.